requirements.txt
matplotlib==3.3.1
torch==1.6.0
torchvision==0.7.0
Pillow==7.2.0
pytorch-transformers==1.2.0
numpy==1.19.4
pandas==1.1.2
tqdm==4.51.0